Consolidated grammar models
