--- src/STANDS4.Client/Models/Grammar/GrammarCategory.cs
@@ -1,4 +1,0 @@
-namespace Stands4.Models
-{
-    public record GrammarCategory(string Id, string Name);
-}
