--- src/STANDS4.Client/Models/Grammar/GrammarCheckModel.cs
@@ -1,13 +1,0 @@
-using System.Collections.Generic;
-
-
-namespace Stands4.Models
-{
-    public record GrammarCheckModel
-    (
-        string Software,
-        GrammarWarning Warnings,
-        GrammarLanguage Language,
-        List<GrammarMatch> Matches
-    ) : IGrammarCheckResult;
-}
