--- src/STANDS4.Client/Models/Grammar/GrammarError.cs
@@ -1,13 +1,0 @@
-namespace Stands4.Models
-{
-    public class GrammarErrorModel
-    {
-        public GrammarError Results { get; init; } = new();
-    }
-
-
-    public class GrammarError
-    {
-        public string Error { get; init; } = string.Empty;
-    }
-}
