--- src/STANDS4.Client/Models/Grammar/GrammarExceptionModel.cs
@@ -1,7 +1,0 @@
-namespace Stands4.Models
-{
-    public record GrammarExceptionModel
-    (
-        string Error
-    ) : IGrammarCheckResult;
-}
