--- src/STANDS4.Client/Models/Grammar/GrammarLanguage.cs
@@ -1,17 +1,0 @@
-namespace Stands4.Models
-{
-    public record GrammarLanguage
-    (
-        string Name,
-        string Code,
-        GrammarLanguageDetected languageDetected
-    );
-
-
-    public record GrammarLanguageDetected
-    (
-        string Name,
-        string Code,
-        decimal Confidence
-    );
-}
