--- src/STANDS4.Client/Models/Grammar/GrammarReplacement.cs
@@ -1,4 +1,0 @@
-namespace Stands4.Models
-{
-    public record GrammarReplacement(string Value, string ShortDescription = "");
-}
