--- src/STANDS4.Client/Models/Grammar/GrammarRule.cs
@@ -1,10 +1,0 @@
-namespace Stands4.Models
-{
-    public record GrammarRule
-    (
-        string Id,
-        string Description,
-        string IssueType,
-        GrammarCategory Category
-    );
-}
