--- src/STANDS4.Client/Models/Grammar/GrammarType.cs
@@ -1,4 +1,0 @@
-namespace Stands4.Models
-{
-    public record GrammarType(string TypeName);
-}
