--- src/STANDS4.Client/Models/Grammar/GrammarWarning.cs
@@ -1,4 +1,0 @@
-namespace Stands4.Models
-{
-    public record GrammarWarning(bool IncompleteResults);
-}
