--- src/STANDS4.Client/Models/Grammar/IGrammarCheckResult.cs
@@ -1,8 +1,0 @@
-namespace Stands4.Models
-{
-    internal interface IGrammarCheckResult
-    {
-        // this interface is intentionally blank
-        // it provies a common base type for objects that can be deserialized by System.Text.Json
-    }
-}
